Removed models from load_once_paths.
Previously, there was a strange problem where associations between
models in the plugin and the app worked fine on an initial request
(after restarting the server), but not for subsequent requests.

A temporary workaround was to replace:

    @issue.account

with the ugly, but functional:

    Account.find(@issue.account_id)

Now that Issue loads account through a join table, this workaround
is no longer suitable.

Removing items from load_once_paths was advised here:
    http://www.spacevatican.org/2008/5/28/reload-me-reload-me-not