Created permissions section for issues form
Found a new problem: the way that Account is being opened up to add
the has_many :issues association is causing the :belongs_to :user
association (and presumably others) to be lost.

next step: write tests to ensure that Account maintains associations
from Fat Free CRM as well as those added in this plugin.